% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_missing_values.R
\name{count_missing_values}
\alias{count_missing_values}
\title{Count missing values for vars (by group)}
\usage{
count_missing_values(df, vars, group = NULL, pivot = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{vars}{Variable names.}

\item{group}{Grouping column names. Default to NULL.}

\item{pivot}{Logical. If TRUE, return a pivoted data frame.}
}
\description{
Count missing values for vars (by group)
}
