% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_dm.R
\name{date_to_dm}
\alias{date_to_dm}
\title{Date to day and month}
\usage{
date_to_dm(vec)
}
\arguments{
\item{vec}{A vector of type POSIXct.}
}
\value{
The same vector with only the day and month in the env locale.
}
\description{
Date to day and month
}
