% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deselect.R
\name{deselect}
\alias{deselect}
\title{Remove columns from a data frame}
\usage{
deselect(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Column names.}
}
\value{
A tibble with columns removed
}
\description{
Remove columns from a data frame
}
