% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_col_start.R
\name{df_col_start}
\alias{df_col_start}
\alias{df_col_end}
\alias{df_col_start_end}
\title{Get colnames that start, end, or both with a string}
\usage{
df_col_start(df, pattern_start)

df_col_end(df, pattern_end)

df_col_start_end(df, pattern_start, pattern_end)
}
\arguments{
\item{df}{A data frame.}

\item{pattern_start}{A start string}

\item{pattern_end}{An end string.}
}
\value{
A vector of column names
}
\description{
Get colnames that start, end, or both with a string
}
