% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_diff.R
\name{df_diff}
\alias{df_diff}
\title{Remove columns from a data frame that exists in another tibble}
\usage{
df_diff(df_a, df_b, ...)
}
\arguments{
\item{df_a}{A data frame to remove columns from.}

\item{df_b}{A data frame to extract columns names from.}

\item{...}{Columns from df_b and df_a to keep in data frame.}
}
\value{
A tibble with some columns removed.
}
\description{
Remove columns from a data frame that exists in another tibble
}
