% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv.R
\name{import_csv}
\alias{import_csv}
\alias{import_xlsx}
\alias{import_full_xlsx}
\title{Initial import with clean names (not an option!)}
\usage{
import_csv(path, delim = ",", clean_names = TRUE, guess_max = 21474836)

import_xlsx(path, sheet = 1, clean_names = TRUE, guess_max = 21474836)

import_full_xlsx(path = NULL, clean_names = TRUE, guess_max = 21474836)
}
\arguments{
\item{path}{The project relative path to the file.}

\item{delim}{Field separator, default is ",".}

\item{clean_names}{Boolean. If TRUE, uses \code{\link[janitor:clean_names]{janitor::clean_names()}} to column clean names.}

\item{guess_max}{Maximum guess of types. Default to 21474836.}

\item{sheet}{A character string of the sheet name or the sheet position}
}
\value{
A dataframe.
}
\description{
Import a dataset from a CSV or a XLSX file. It ensures that the right type of column is guessed using the max value of guess_max by default.
}
\section{Functions}{
\itemize{
\item \code{import_csv()}: Import a .csv file.

\item \code{import_xlsx()}: Import one sheet from a .xlsx.

\item \code{import_full_xlsx()}: Import all sheets from a .xlsx.

}}
