% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_dup.R
\name{left_join_dup}
\alias{left_join_dup}
\title{Left join a list of dataframes}
\usage{
left_join_dup(df_a, df_b, ..., duplicated_columns = 3)
}
\arguments{
\item{df_a}{A dataframe.}

\item{df_b}{Another dataframe.}

\item{...}{Columns to join by.}

\item{duplicated_columns}{Strategy between 1 and 4. Default to 3. See section "Treating duplicated columns.}
}
\value{
A left-joined data frame.
}
\description{
Left join a list of dataframes
}
\section{Treating duplicated columns}{


There are 4 ways implemented on treating columns with the same names in both datasets:
\itemize{
\item 1: an usual left-join with suffix being empty for df_a and ".dup_col" for df_b;
\item 2: removing all duplicated columns from df_b, keeping only those of df_a;
\item 3: remove the column from df_b if all values from df_b that are not NAs equal the values from df_a;
\item 4: remove the column from df_b if all values from df_b are identical to values from df_a.
}
}

