% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_joints.R
\name{left_joints}
\alias{left_joints}
\title{Left join a few data frames in a list}
\usage{
left_joints(list, ...)
}
\arguments{
\item{list}{A list of data frames.}

\item{...}{Grouping columns.}
}
\value{
A left-joined data frame.
}
\description{
Left join a few data frames in a list
}
