% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_if_nulla.R
\name{mutate_if_nulla}
\alias{mutate_if_nulla}
\title{Mutate with replacement if NULL or NA values}
\usage{
mutate_if_nulla(df, col, replacement)
}
\arguments{
\item{df}{A data frame.}

\item{col}{A column of \code{df}.}

\item{replacement}{A replacement of the same type as \code{col}.}
}
\value{
A mutated data frame.
}
\description{
Mutate with replacement if NULL or NA values
}
