% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_group_split.R
\name{named_group_split}
\alias{named_group_split}
\title{Named group split}
\usage{
named_group_split(df, group)
}
\arguments{
\item{df}{A data frame.}

\item{group}{Quoted column name to group by.}
}
\value{
A split and named list of data frames.
}
\description{
Named group split
}
