% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_values.R
\name{recode_values}
\alias{recode_values}
\alias{recode_na}
\title{Replace values of a data frame's columns to one value}
\usage{
recode_values(df, values, to_value, ...)

recode_na(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{values}{A vector of values.}

\item{to_value}{A value.}

\item{...}{Column names.}
}
\value{
A data frame with recoded values to one value.
}
\description{
\code{recode_values()} recodes a singular value. \code{recode_na()} recodes usual values of NAs, for instance 999, NULL, NaN, and NA.
}
\details{
If the column type is a character and the replacement a numeric, then the numeric is coerced to a character. If the column type is a numeric and the replacement is a character, then the column is coerced to character. NAs will remains NAs of the right type.
}
