% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_optimum.R
\name{row_optimum}
\alias{row_optimum}
\title{Row-wise optimum (default to pmax) of several columns}
\usage{
row_optimum(
  df,
  ...,
  optimum = "max",
  max_name = "pmax",
  min_name = "pmin",
  na_rm = TRUE,
  keep = "all"
)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Numeric column names.}

\item{optimum}{Should we calculate "max", "min" or "both"? Default to TRUE.}

\item{max_name}{Column name for the mutated pmax.}

\item{min_name}{Column name for the mutated pmin.}

\item{na_rm}{Remove NAs. Default to TRUE.}

\item{keep}{To be used within mutate. Default to "all".}
}
\value{
A data frame with pmax, pmin or both (and all columns or none, depending on "keep")
}
\description{
Row-wise optimum (default to pmax) of several columns
}
