% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_string_count.R
\name{row_string_count}
\alias{row_string_count}
\alias{row_na_count}
\title{Count the number of row-wise occurrences of a pattern}
\usage{
row_string_count(df, pattern = "", new_colname = "count")

row_na_count(df, new_colname = "count")
}
\arguments{
\item{df}{A data frame.}

\item{pattern}{A pattern to pass to \code{stringr::str_count()}. Default to "dnk".}

\item{new_colname}{The newly-mutated column name. Default to "count".}
}
\value{
A mutated data frame.
}
\description{
\code{row_pattern_count()} count row-wise occurences of a pattern over the data frame; \code{row_na_count()} count row-wise number of NAs over the data frame.
}
