% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subvec_in.R
\name{subvec_in}
\alias{subvec_in}
\alias{subvec_not_in}
\title{Get a subset of a vector}
\usage{
subvec_in(vector, set)

subvec_not_in(vector, set)
}
\arguments{
\item{vector}{A vector to subset.}

\item{set}{A set-vector.}
}
\value{
A subset of a list or a vector.
}
\description{
\code{subvec_in()} gets a subset of a vector if vector elements are in a given \code{set}. \code{subvec_not_in()} get a subset of a vector if vector elements are NOT in a fiven \code{set}.
}
