% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_joints.R
\name{left_joints}
\alias{left_joints}
\title{Left join a list of dataframes}
\usage{
left_joints(list, ...)
}
\arguments{
\item{list}{A list of data frames.}

\item{...}{Columns to join by.}
}
\value{
A left-joined data frame.
}
\description{
Left join a list of dataframes
}
