% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_joints.R
\name{left_joints}
\alias{left_joints}
\alias{left_joints_dup}
\title{Left join a list of dataframes}
\usage{
left_joints(list, ...)

left_joints_dup(list, ..., duplicated_columns = 3)
}
\arguments{
\item{list}{A list of data frames.}

\item{...}{Columns to join by.}

\item{duplicated_columns}{Strategy between 1 and 4. Default to 3. See section "Treating duplicated columns.}
}
\value{
A left-joined data frame.
}
\description{
\link{left_joints} is a simple reducing functions based on \link[dplyr:mutate-joins]{dplyr::left_join}. \link{left_joints_dup} is a reduce wrapper around \link{left_join_dup}.
}
